quell cast size mismatch warning on 64b builds
svn:revision:30133
svn:branch:trunk
svn:account:erikgreenwald

+++ src/vdeck/vdeck.c
@@ -405,7 +405,7 @@ flatten_tree( struct bu_vls *vls, union tree *tp, char *op, int neg )
 		return;
 
 	case OP_REGION:
-		bu_log("REGION 'stp'=x%x\n", (unsigned int)tp->tr_a.tu_stp );
+		bu_log("REGION 'stp'=x%x\n", (size_t)tp->tr_a.tu_stp );
 		return;
 
 	default:
